// FrontEnd/src/types/models/Situation.ts
export interface Situation {
    id_situation: number;
    category: string;
    first_choice: string;
    first_choice_electric_charge: number;
    question: string;
    second_choice: string;
    second_choice_electric_charge: number;
    third_choice: string;
    third_choice_electric_charge: number;
    right_answer: string;
    advise: string;
  }